generative-ui/personal-assistant/requirements.txt
autogen>=0.4.2
beautifulsoup4
requests
selenium
webdriver_manager
python-dotenv 